% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{clean.tweets}
\alias{clean.tweets}
\title{Performs all necessary cleaning on a data frame of tweets. This includes removing all symbols from tweets, converting
them to a lower case array of words, removing all stop words from this array, andconverting timestamps to an
R usable format. Can also filter by time zone if desired (default does not filter)}
\usage{
clean.tweets(tweets.df, tz = NULL, stoplist = NULL)
}
\arguments{
\item{tweets.df}{An array of tweets with desired variables attached. (Use dplyr to filter variables)}

\item{tz}{A list of time zones to filter by, currently case sensitive}

\item{stoplist}{The stoplist used to filter words}
}
\value{
The tweet data frame with all editing / filtering done.
}
\description{
Performs all necessary cleaning on a data frame of tweets. This includes removing all symbols from tweets, converting
them to a lower case array of words, removing all stop words from this array, andconverting timestamps to an
R usable format. Can also filter by time zone if desired (default does not filter)
}
\examples{
\dontrun{df = select(rawdata, text, time_zone)}
\dontrun{tweets = clean.tweets(dataframe)}
\dontrun{tweets = clean.tweets(dataframe, tz = c("Pacific Time (US & Canada)", "Eastern Time (US & Canada)),
stoplist = stoplist))}
}

