% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{clean_dates}
\alias{clean_dates}
\title{Formats timestamps}
\usage{
clean_dates(x)
}
\arguments{
\item{x}{The date of a tweet}
}
\value{
The date in the proper format to convert to a posixt format
}
\description{
Remove extra information from dates and sort them
}
\examples{
\dontrun{date = clean_dates(date)}
}

