% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{clean_links}
\alias{clean_links}
\title{Handles link symbols}
\usage{
clean_links(x)
}
\arguments{
\item{x}{A string or array of strings to remove symbols from}
}
\value{
The string or array of strings with the symbols removed
}
\description{
Removes all punctuation from tweet text while also removing the protocol (http) and domain (.com) indicators from
any links in the text.
}
\examples{
\dontrun{text = unlist(lapply(tweets$text, clean_links))}
}

