% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{clean_text}
\alias{clean_text}
\title{Text cleaner}
\usage{
clean_text(x)
}
\arguments{
\item{x}{A string to remove symbols from}
}
\value{
The string with the symbols removed
}
\description{
Cleans all garbage symbols, links, retweet tags, etc from the tweet text.
}
\examples{
\dontrun{text = unlist(lapply(tweet.df$text, clean_text))}
}

