% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\docType{data}
\name{fergusontweets}
\alias{fergusontweets}
\title{5634 Tweets about Ferguson}
\format{A data frame with 5634 rows and 42 columns}
\source{
Downloaded using the streamR filterstream function.
}
\usage{
fergusontweets
}
\description{
A dataset containing the text and metadata of 5634 tweets containing keywords concerning Ferguson or other
related incidents.
}
\keyword{datasets}

