% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{filter_html}
\alias{filter_html}
\title{Removes all html codes that represent symbols / puncation.}
\usage{
filter_html(x)
}
\arguments{
\item{x}{A string or array of strings to remove symbols from}
}
\value{
The string or array of strings with the symbols removed
}
\description{
Removes all html codes that represent symbols / puncation.
}
\examples{
\dontrun{text = filter_html(text)}
}

