% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{locate_tweets}
\alias{locate_tweets}
\title{Matches tweets to zipcodes}
\usage{
locate_tweets(located_tweets)
}
\arguments{
\item{located_tweets}{A cleaned array of tweets with latitude / longitude variables.}
}
\value{
The tweet data frame with all in US tweets labeled with a state and all others with NA
}
\description{
Labels an array of geo-located tweets with state abbreviations using shapefiles provided by the Census Bureau. This function
should be run after cleaning the tweets, and only works on tweets with a latitude / longitude. Any tweets without
lattitude / longitude codes are labeled with NA.
}
\examples{
\dontrun{tweets = clean_tweets(dataframe)}
\dontrun{located_tweets = locate_tweets(tweets)}
}

