% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\docType{data}
\name{locatedtweets}
\alias{locatedtweets}
\title{7585 Tweets from the US}
\format{A data frame with 7585 rows and 42 columns}
\source{
Downloaded using the streamR filterstream function.
}
\usage{
locatedtweets
}
\description{
A dataset containing the text and metadata of 7585 tweets on any topic sent from the US.
}
\keyword{datasets}

