% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{mywhich}
\alias{mywhich}
\title{Handles stopwords}
\usage{
mywhich(word.vector, stoplist)
}
\arguments{
\item{word.vector}{An array of words to filter}

\item{stoplist}{The stoplist to filter with}
}
\value{
The word array with all stopwords removed.
}
\description{
Removes all words in an array that are on a given stoplist
}
\examples{
\dontrun{text = strsplit(text, " ")}
\dontrun{text = lapply(text, mywhich, stoplist)}
}

