% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{process.files}
\alias{process.files}
\title{Processes all tweet-files in a directory and saves them in the given directory. This functions filters each
files variables, runs clean.tweets on them, and runs locate.tweets on them if desired.}
\usage{
process.files(tweetdir, outputdir, loc = FALSE, vars = "text", ...)
}
\arguments{
\item{tweetdir}{The file directory of tweets to read from. No files other than the tweet files should be in this
directory. The tweet files should be in the format as provided by streamR's parseTweets.}

\item{outputdir}{The location to save the tweets to.}

\item{loc}{Determines if the tweets should be zip code tagged. Boolean variable}

\item{vars}{The variables to filter from the tweet files.}

\item{...}{tz and stoplist arguments for the inner clean.tweets call. Defaults both to false}
}
\value{
The list of filenames edited.
}
\description{
Processes all tweet-files in a directory and saves them in the given directory. This functions filters each
files variables, runs clean.tweets on them, and runs locate.tweets on them if desired.
}
\examples{
\dontrun{process.files("~/Documents/RawTweets", "~/Documents/EditedTweets")}
\dontrun{process.files("~/Documents/RawTweets", "~/Documents/EditedTweets", loc = TRUE,
vars = c("text", "time_zone"), tz = c("Jerusalem", "NA"), stoplist = stoplist)}
}

