% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{process_files}
\alias{process_files}
\title{Processes a folder of tweet files}
\usage{
process_files(tweetdir, outputdir, loc = FALSE, vars = "text",
  makedf = TRUE, ...)
}
\arguments{
\item{tweetdir}{The file directory of tweets to read from. No files other than the tweet files should be in this
directory. The tweet files should be in the format as provided by streamR's parseTweets.}

\item{outputdir}{The location to save the tweets to.}

\item{loc}{True if the tweets should be tagged with zip codes. Uses 2014 shapefile from census.gov}

\item{vars}{The variables to filter from the tweet files.}

\item{makedf}{True if a dataframe containing all the edited tweets should be returned, false otherwise. Defaults
to true.}

\item{...}{tz and stoplist arguments for the inner clean.tweets call.}
}
\value{
The list of filenames edited.
}
\description{
Processes all tweet-files in a directory and saves them in the given directory. This functions filters each
files variables, removes symbols, emotes and stopwords, and tags tweets with zip codes if desired.
}
\examples{
\dontrun{process_files("~/Documents/RawTweets", "~/Documents/EditedTweets")}
\dontrun{process_files("~/Documents/RawTweets", "~/Documents/EditedTweets", loc = TRUE,
vars = c("text", "time_zone"), tz = c("Eastern Time (US & Canada)", "NA"), stoplist = stoplist)}
}

