% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\name{trim}
\alias{trim}
\title{Handles leading / trailing white space.}
\usage{
trim(x)
}
\arguments{
\item{x}{The tweet to remove white space from.}
}
\value{
The string or array of strings with the white space removed
}
\description{
Removes all trailing / leading white space from a tweet.
}
\examples{
\dontrun{text = trim(text)}
}

