% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTweetFiles.R
\docType{data}
\name{unlocatedtweets}
\alias{unlocatedtweets}
\title{655 random tweets}
\format{A data frame with 655 rows and 42 columns}
\source{
Downloaded using the streamR filterstream function.
}
\usage{
unlocatedtweets
}
\description{
A dataset containing the text and metadata of 655 on any topic sent from any location.
}
\keyword{datasets}

